packages/agentkit-builtin/obs-otel/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentkit-builtin-obs-otel"
version = "0.1.0"
description = "OpenTelemetry observability technique for AgentKit."
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "agentkit>=0.1.0",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-http>=1.27",
]

[project.entry-points."agentkit.techniques"]
"observability.otel" = "agentkit_builtin_obs_otel:OtelObservability"

[tool.hatch.build.targets.wheel]
packages = ["src/agentkit_builtin_obs_otel"]
